REF: Move extract_title into shared utils module.
