+++ web_monitoring/cli.py
@@ -4,7 +4,6 @@
 from docopt import docopt
 import hashlib
 import io
-import lxml.html
 import requests
 from tqdm import tqdm
 from web_monitoring import internetarchive as ia
@@ -16,13 +15,6 @@ from web_monitoring import db
 # better to use the underlying library code.
 
 
-def _extract_title(content_bytes):
-    "Return content of <title> tag as string."
-    content_as_file = io.StringIO(content.decode(errors='ignore'))
-    title = lxml.html.parse(content_as_file).find(".//title").text
-    return title
-
-
 def import_ia(url, agency, site):
     print('obtaining versions list from Internet Archive...')
     versions = ia.list_versions(url)
@@ -32,7 +24,7 @@ def import_ia(url, agency, site):
         res = requests.get(uri)
         assert res.ok
         version_hash = hashlib.sha256(res.content).digest()
-        title = _extract_title(res.content)
+        title = extract_title(res.content)
         v = ia.format_version(url=url, dt=dt, uri=uri,
                               version_hash=version_hash, title=title,
                               agency=agency, site=site)
