+++ web_monitoring/utils.py
@@ -1,0 +1,8 @@
+import lxml.html
+
+
+def _extract_title(content_bytes):
+    "Return content of <title> tag as string."
+    content_as_file = io.StringIO(content.decode(errors='ignore'))
+    title = lxml.html.parse(content_as_file).find(".//title").text
+    return title
